Refactor: divide app into self contained components
